Move `queryBy*` functions from the `LSPTask` to a separate file
