+++ main/lsp/AbstractRenamer.cc
@@ -1,5 +1,7 @@
 #include "AbstractRenamer.h"
+#include "main/lsp/LSPQuery.h"
 #include "main/lsp/lsp.h"
+
 using namespace std;
 
 namespace sorbet::realmain::lsp {
